bayes: need ability to have nodes to have nonzero prob of non-baseline vals
Without this, it's way too easy to get a set of observations that has zero
probability assigned to it. (e.g. as it stands, RR can't be high without
one of the hemorrhage nodes being true)

First step: make-naive-assumptions.py reads priors from scenario-bn.yaml, but
all priors are probability 1.0 of the baseline.

Verified that this gives the same results +/- floating point error.